# Use hardware accelerated CRC32C function if available