--- a/tox.ini
+++ b/tox.ini
@@ -18,6 +18,7 @@ deps =
     python-snappy
     lz4
     xxhash
+    crc32c
     py26: unittest2
 commands =
     py.test {posargs:--pylint --pylint-rcfile=pylint.rc --pylint-error-types=EF --cov=kafka --cov-config=.covrc}