@dpkp @jeffwidman Hehe, I'm an idiot for this =) So, here I was pushing my butt to accelerate varint implementation, but did not look properly at the profiler.
Long story short, while accelerating the V2 implementation for `aiokafka` I found that `_crc32c.py` is super slow and takes more time for 1 call than 100 `append`. So as a fast fix I propose to use https://github.com/ICRAR/crc32c if available on the system. I will look into `_crc32c.py` script later to make it a bit faster.

Benchmark without HW crc32:
```
.....................
batch_read_v0: Mean +- std dev: 1.42 ms +- 0.03 ms
.....................
batch_read_v1: Mean +- std dev: 1.44 ms +- 0.03 ms
.....................
batch_read_v2: Mean +- std dev: 3.10 ms +- 0.06 ms
```

With HW crc32:
```
.....................
batch_read_v0: Mean +- std dev: 1.48 ms +- 0.05 ms
.....................
batch_read_v1: Mean +- std dev: 1.49 ms +- 0.05 ms
.....................
batch_read_v2: Mean +- std dev: 1.09 ms +- 0.04 ms
```

So now V2 is relatively the same speed as V1 and V0. Now we can move on to TransactionalProducer =)