@jeffwidman 2018-02-21
Oh interesting. Two thoughts: 
1. if we go down this route of fixing by adding an optional dependency, there should be a note about it in the docs as an optional way to speedup `kafka-python`.
2. It's not clear to me what the other options are beyond adding an optional dependency... since this isn't pure python, does that create problems with directly vendoring it in like we do a few other libs?