@tvoinarovskyi 2018-02-21
@jeffwidman Yea, will patch the docs.
Yes, you can't easiely bundle a C extension module, you will need to correctly build it. As for alternitives, in aiokafka I plan to use the C implementation taken from librdkafka...
I will also try to make the python script at least somewhat faster.